pyproject.toml
[tool.poetry]
name = "permissions-service"
version = "0.1.0"
description = "Serverless permissions CRUD service"
authors = ["Yishai Wiesner <Yishai.Wiesner@cyberark.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
boto3 = "1.40.4"
PyJWT = "~2.12.1"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "~9.0.2"
moto = {extras = ["dynamodb", "sqs"], version = "^5.1.22"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
